rename test_util and add comments
