+++ relengapi/blueprints/tokenauth/test_tokenauth.py
@@ -12,12 +12,12 @@ from nose.tools import eq_
 from relengapi import p
 from relengapi.blueprints.tokenauth import types
 from relengapi.blueprints.tokenauth.tables import Token
-from relengapi.blueprints.tokenauth.test_util import FakeSerializer
-from relengapi.blueprints.tokenauth.test_util import insert_all
-from relengapi.blueprints.tokenauth.test_util import insert_prm
-from relengapi.blueprints.tokenauth.test_util import insert_usr
-from relengapi.blueprints.tokenauth.test_util import prm_json
-from relengapi.blueprints.tokenauth.test_util import usr_json
+from relengapi.blueprints.tokenauth.util import FakeSerializer
+from relengapi.blueprints.tokenauth.util import insert_all
+from relengapi.blueprints.tokenauth.util import insert_prm
+from relengapi.blueprints.tokenauth.util import insert_usr
+from relengapi.blueprints.tokenauth.util import prm_json
+from relengapi.blueprints.tokenauth.util import usr_json
 from relengapi.lib import auth
 from relengapi.lib.testing.context import TestContext
 from wsme.rest.json import fromjson
